+++ tests/unit/test_client.py
@@ -296,11 +296,33 @@ class TestClient:
 
         optimeout_class.assert_called()
 
+    @pytest.mark.parametrize("command", ["gat", "gats"])
+    async def test_get_and_touch_command_use_timeout(self, client, command, mocker):
+        optimeout_class = mocker.patch("emcache.client.OpTimeout", MagicMock())
+
+        connection = AsyncMock()
+        connection.get_and_touch_command = AsyncMock(return_value=iter([[b"foo"], [b"value"], [0], [0]]))
+        connection_context = AsyncMock()
+        connection_context.__aenter__.return_value = connection
+        node = Mock()
+        node.connection.return_value = connection_context
+        client._cluster.pick_node.return_value = node
+        f = getattr(client, command)
+        await f(0, b"foo")
+
+        optimeout_class.assert_called()
+
     @pytest.mark.parametrize("command", ["get", "gets"])
     async def test_fetch_command_invalid_key(self, client, command):
         with pytest.raises(ValueError):
             f = getattr(client, command)
             await f(b"\n")
+
+    @pytest.mark.parametrize("command", ["gat", "gats"])
+    async def test_get_and_touch_command_invalid_key(self, client, command):
+        with pytest.raises(ValueError):
+            f = getattr(client, command)
+            await f(0, b"\n")
 
     @pytest.mark.parametrize("command", ["get", "gets"])
     async def test_fetch_command_client_closed(self, client, command):
@@ -308,6 +330,13 @@ class TestClient:
         with pytest.raises(RuntimeError):
             f = getattr(client, command)
             await f(b"key")
+
+    @pytest.mark.parametrize("command", ["gat", "gats"])
+    async def test_get_and_touch_command_client_closed(self, client, command):
+        await client.close()
+        with pytest.raises(RuntimeError):
+            f = getattr(client, command)
+            await f(0, b"key")
 
     @pytest.mark.parametrize("command", ["get_many", "gets_many"])
     async def test_fetch_many_command_use_timeout(self, client, command, mocker):
@@ -325,10 +354,32 @@ class TestClient:
 
         optimeout_class.assert_called()
 
+    @pytest.mark.parametrize("command", ["gat_many", "gats_many"])
+    async def test_get_and_touch_many_command_use_timeout(self, client, command, mocker):
+        optimeout_class = mocker.patch("emcache.client.OpTimeout", MagicMock())
+
+        connection = AsyncMock()
+        connection.get_and_touch_command = AsyncMock(return_value=iter([[b"foo"], [b"value"], [0], [0]]))
+        connection_context = AsyncMock()
+        connection_context.__aenter__.return_value = connection
+        node = Mock()
+        node.connection.return_value = connection_context
+        client._cluster.pick_nodes.return_value = {node: [b"foo"]}
+        f = getattr(client, command)
+        await f(0, [b"foo"])
+
+        optimeout_class.assert_called()
+
     @pytest.mark.parametrize("command", ["get_many", "gets_many"])
     async def test_fetch_many_command_empty_keys(self, client, command):
         f = getattr(client, command)
         result = await f([])
+        assert result == {}
+
+    @pytest.mark.parametrize("command", ["gat_many", "gats_many"])
+    async def test_get_and_touch_many_command_empty_keys(self, client, command):
+        f = getattr(client, command)
+        result = await f(0, [])
         assert result == {}
 
     @pytest.mark.parametrize("command", ["get_many", "gets_many"])
@@ -337,12 +388,25 @@ class TestClient:
             f = getattr(client, command)
             await f([b"\n"])
 
+    @pytest.mark.parametrize("command", ["gat_many", "gats_many"])
+    async def test_get_and_touch_many_command_invalid_keys(self, client, command):
+        with pytest.raises(ValueError):
+            f = getattr(client, command)
+            await f(0, [b"\n"])
+
     @pytest.mark.parametrize("command", ["get_many", "gets_many"])
     async def test_fetch_many_command_client_closed(self, client, command):
         await client.close()
         with pytest.raises(RuntimeError):
             f = getattr(client, command)
             await f([b"key"])
+
+    @pytest.mark.parametrize("command", ["gat_many", "gats_many"])
+    async def test_get_and_touch_many_command_client_closed(self, client, command):
+        await client.close()
+        with pytest.raises(RuntimeError):
+            f = getattr(client, command)
+            await f(0, [b"key"])
 
     @pytest.mark.parametrize("command", ["increment", "decrement"])
     async def test_incr_decr_use_timeout(self, client, command, mocker):
@@ -479,7 +543,7 @@ class TestClient:
             await client.flush_all(memcached_host_address)
 
     @pytest.mark.parametrize("command", ["get_many", "gets_many"])
-    async def test_exception_cancels_everything(self, client, command):
+    async def test_exception_cancels_for_fetch_many(self, client, command):
         # patch what is necesary for rasing an exception for the first query and
         # a "valid" response from the others
         connection = AsyncMock()
@@ -496,6 +560,25 @@ class TestClient:
         with pytest.raises(OSError):
             f = getattr(client, command)
             await f([b"key1", b"key2", b"key3"])
+
+    @pytest.mark.parametrize("command", ["gat_many", "gats_many"])
+    async def test_exception_cancels_for_get_and_touch_many(self, client, command):
+        # patch what is necesary for rasing an exception for the first query and
+        # a "valid" response from the others
+        connection = AsyncMock()
+        connection.get_and_touch_command.side_effect = MagicMock(side_effect=iter([OSError(), b"Ok", b"Ok"]))
+        connection_context = AsyncMock()
+        connection_context.__aenter__.return_value = connection
+        node1 = Mock()
+        node2 = Mock()
+        node3 = Mock()
+        node1.connection.return_value = connection_context
+        node2.connection.return_value = connection_context
+        node3.connection.return_value = connection_context
+        client._cluster.pick_nodes.return_value = {node1: [b"key1"], node2: [b"key2"], node3: [b"key3"]}
+        with pytest.raises(OSError):
+            f = getattr(client, command)
+            await f(0, [b"key1", b"key2", b"key3"])
 
     @pytest.mark.parametrize("command", ["append", "prepend"])
     async def test_exptime_flags_disabled(self, client, command):
@@ -576,6 +659,23 @@ class TestClient:
         get_flags.execute.assert_not_called()
         get_noflags.execute.assert_not_called()
         gets_noflags.execute.assert_not_called()
+
+    async def test_version_client_closed(self, client, memcached_host_address):
+        await client.close()
+        with pytest.raises(RuntimeError):
+            await client.version(memcached_host_address)
+
+    async def test_version_error_command(self, client, memcached_host_address):
+        # patch what is necesary for returnning an error string
+        connection = AsyncMock()
+        connection.version_command = AsyncMock(return_value=b"ERROR")
+        connection_context = AsyncMock()
+        connection_context.__aenter__.return_value = connection
+        node = Mock()
+        node.connection.return_value = connection_context
+        client._cluster.node.return_value = node
+        with pytest.raises(CommandError):
+            await client.version(memcached_host_address)
 
 
 async def test_create_client_default_values(event_loop, mocker):